refactor: layout -> page로 page -> 컴포넌트로 이동
